Add in maxReadBuffer, fix http read
maxReadBuffer is now used in filesystem and internet
Verify socket.read or http.read have a valid number
Make http.read use 'n' argument
Add in inf to config's loading environment so infinity can be stored.

The reason that http used to ignore read was because specifying a number
used to be ignored, it turns out that OC was reading from the wrong
index, 2 instead of 1, causing it to be ignored.