Add in beep for all platforms
Add client.beepSampleRate and client.beepVolume to the config
Add binding for SDL_QueueAudio
Add SDL_QueueAudio based beep system

This is very minimal. It generates all of the audio at once and queues
the entire thing (which OpenComputers itself does as well) instead of
gradually generating and streaming audio.
Perhaps in a later commit.